refactor: split linkup_server into modules
