+++ server-tests/tests/helpers.rs
@@ -1,7 +1,8 @@
 use std::process::Command;
 
 use linkup::{Domain, MemoryStringStore, SessionService, UpdateSessionRequest};
-use linkup_local_server::{linkup_router, DnsCatalog};
+
+use linkup_local_server::{router, DnsCatalog};
 use reqwest::Url;
 use tokio::net::TcpListener;
 
@@ -14,7 +15,7 @@ pub enum ServerKind {
 pub async fn setup_server(kind: ServerKind) -> String {
     match kind {
         ServerKind::Local => {
-            let app = linkup_router(MemoryStringStore::default(), DnsCatalog::new());
+            let app = router(MemoryStringStore::default(), DnsCatalog::new());
 
             // Bind to a random port assigned by the OS
             let listener = TcpListener::bind("127.0.0.1:0").await.unwrap();
